🔨 - simplify generator code + update dependencies
