--- .babelrc
@@ -1,4 +1,0 @@
-{
-  "plugins": ["babel-plugin-add-module-exports"],
-  "presets": ["es2015"]
-}
