+++ .gitignore
@@ -1,5 +1,3 @@
 node_modules
 npm-debug.log
 test/temp/
-
-*.sublime-workspace
