+++ .npmignore
@@ -6,5 +6,3 @@ npm-debug.log
 .eslintrc
 .gitignore
 .travis.yml
-.sublime-workspace
-.sublime-project
