+++ package.json
@@ -16,10 +16,7 @@
     "node": ">=4.2.2"
   },
   "scripts": {
-    "build": "babel src/generator.js --out-file app/index.js",
-    "watch": "babel src/generator.js --watch --out-file app/index.js",
-    "prepublish": "npm run build",
-    "lint": "eslint src/generator.js",
+    "lint": "eslint app/index.js",
     "test": "ava test/test.js --fail-fast --verbose"
   },
   "files": [
@@ -36,24 +33,19 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "chalk": "^2.1.0",
-    "globby": "^6.0.0",
+    "chalk": "^2.3.0",
     "lodash": "^4.16.4",
     "mkdirp": "^0.5.1",
-    "superb": "^1.3.0",
-    "yeoman-generator": "^0.24.1",
+    "superb": "^2.0.0",
+    "yeoman-generator": "^2.0.1",
     "yosay": "^2.0.1"
   },
   "devDependencies": {
-    "ava": "^0.16.0",
-    "babel-cli": "^6.16.0",
-    "babel-core": "^6.17.0",
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-preset-es2015": "^6.16.0",
-    "eslint": "^4.6.1",
+    "ava": "^0.23.0",
+    "eslint": "^4.11.0",
     "eslint-config-kraftvaerk": "0.0.5",
     "pify": "^3.0.0",
-    "yeoman-assert": "^3.0.0",
+    "yeoman-assert": "^3.1.0",
     "yeoman-test": "^1.5.1"
   }
 }
